requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
django-allauth==0.40.0
django-bootstrap3==11.1.0
django-ckeditor==5.7.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-model-utils==3.2.0
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-taggit==1.1.0
gunicorn==19.9.0
idna==2.8
jsonfield==2.0.2
Markdown==3.1.1
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
Pygments==2.4.2
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.6
whitenoise==4.1.4
